+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/IEnchantmentTooltip.java
@@ -1,8 +1,0 @@
-package insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment;
-
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.ItemStack;
-
-public interface IEnchantmentTooltip {
-    Component getTooltip(ItemStack stack, int lvl);
-}